// Arashyn-FE/src/lib/api/generated/models/deck-summarise-response.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Arashyn API
 * REST API for Arashyn Platform
 *
 * The version of the OpenAPI document: v1.0
 * Contact: touman876@gmail.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



export interface DeckSummariseResponse {
    'id'?: string;
    'name'?: string;
    'description'?: string;
    'language'?: DeckSummariseResponseLanguageEnum;
    'ownerId'?: string;
    'ownerName'?: string;
}

export enum DeckSummariseResponseLanguageEnum {
    Vi = 'VI',
    En = 'EN',
    Ja = 'JA',
    Ko = 'KO',
    Zh = 'ZH'
}